--- a/osu.Framework.iOS/IOSWindow.cs
+++ b/osu.Framework.iOS/IOSWindow.cs
@@ -51,6 +51,14 @@ namespace osu.Framework.iOS
 
         protected override void RunMainLoop()
         {
+            // Delegate running the main loop to CADisplayLink.
+            //
+            // Note that this is most effective in single thread mode.
+            // .. In multi-threaded mode it will time the *input* thread to the callbacks. This is kinda silly,
+            // but users shouldn't be using multi-threaded mode in the first place. Disabling it completely on
+            // iOS may be a good forward direction if this ever comes up, as a user may see a potentially higher
+            // frame rate with multi-threaded mode turned on, but it is going to give them worse input latency
+            // and higher power usage.
             SDL.SDL_iPhoneSetEventPump(SDL.SDL_bool.SDL_FALSE);
             SDL.SDL_iPhoneSetAnimationCallback(SDLWindowHandle, 1, _ => RunFrame(), IntPtr.Zero);
         }