Alternative to and closes https://github.com/ppy/osu-framework/pull/5803 (the iOS portion – the macOS part is still
valid but only requires a veldrid version bump).

Testing at 2x limiter, 120hz display (iPad Pro 2018), release build local framework

| Header      | menu (contracted) | menu (expanded) | song select |
|-------------|-------------------|-----------------|-------------|
| `master`    | 120fps                 | 70fps               | 28fps           |
| this PR     | 120fps            | 80fps           | 28fps       |
| `master` gl | 120fps                 | 70fps               | 27fps           | 

Not much difference, but that's because we're still doing way too much work in draw. The main thing I want to show is that nothing regressed. Once we have more overhead this is only going to help more.

Of note, in a very simple framework visual test I *can* feel a reduction in input latency on this PR, similar to what I reported on macOS.